moved AbstractMapFeature
