+++ android/rctmgl/src/main/java-v10/com/mapbox/rctmgl/components/styles/sources/AbstractSourceConsumer.kt
@@ -1,7 +1,7 @@
 package com.mapbox.rctmgl.components.styles.sources
 
 import android.content.Context
-import com.mapbox.rctmgl.components.camera.AbstractMapFeature
+import com.mapbox.rctmgl.components.AbstractMapFeature
 
 abstract class AbstractSourceConsumer(context: Context?) : AbstractMapFeature(context) {
     abstract val iD: String?
